Extended gui
This gui still does not work. I changed the frame structure and created
new sections. I also added a lot of comments. The looks of the "Deck Selection" and "Initialize Calculation" section are more ore less finished. Their functionsmust be made compatible with the calculation modules in a next step. I also
started with the pool section of the gui, but it has not been nearly
finished yet.